food_order now saves food orders

+++ app/src/main/java/com/scottb4/lunchilicious/data/FoodOrderEntityClass.kt
@@ -5,7 +5,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "food_order")
 data class FoodOrder (
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val total_cost: Double
 )
